requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.5
Django==5.2.12
django-modeltranslation==0.20.2
django-rosetta==0.10.3
gunicorn==25.1.0
idna==3.11
packaging==26.0
pillow==12.1.1
polib==1.2.0
python-dotenv==1.2.2
requests==2.32.5
sqlparse==0.5.5
typing_extensions==4.15.0
urllib3==2.6.3
whitenoise==6.12.0
